requirements.txt
praw~=7.2.0
pymongo~=3.11.4
jsonpickle~=2.0.0
dnspython~=2.1.0
python-dotenv~=0.17.1